solve delete grade_item issue

+++ Modules/QuestionBank/Http/Controllers/QuizController.php
@@ -312,7 +312,8 @@ class QuizController extends Controller
 
         $quiz=Quiz::where('id', $request->quiz_id)->first();
         $grade_item=GradeItems::where('item_Entity',$request->quiz_id)->first();
-        $grade_item->delete();
+        if(isset($grade_item))
+            $grade_item->delete();
         $quiz->delete();
         return HelperController::api_response_format(200, [], 'Quiz deleted successfully');
     }
